+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -5,7 +5,9 @@ CHANGELOG
 ---
 
  * `TokenInterface` does not extend `Serializable` anymore
- *  Remove all classes in the `Core\Encoder\`  sub-namespace, use the `PasswordHasher` component instead
+ * Remove all classes in the `Core\Encoder\`  sub-namespace, use the `PasswordHasher` component instead
+ * Remove methods `getPassword()` and `getSalt()` from `UserInterface`, use `PasswordAuthenticatedUserInterface`
+   or `LegacyPasswordAuthenticatedUserInterface` instead
 
 5.3
 ---